% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_fragment_file.R
\name{AggregateFragmentFile}
\alias{AggregateFragmentFile}
\title{AggregateFragmentFile parallel}
\usage{
AggregateFragmentFile(
  input_file,
  membership,
  output_name = NULL,
  output_path = NULL,
  tmp_path = "./tmp/",
  nb_cl = NULL,
  nb_row_split = "10000000",
  bgzip_path = NULL,
  tabix_path = NULL,
  returnOutputFileName = T
)
}
\arguments{
\item{input_file}{path to original fragments file}

\item{membership}{membership vector}

\item{output_name}{name of the supercell fragments file (whithout extension)}

\item{output_path}{path where the new fragments file will be saved}

\item{tmp_path}{directory path for tempory files}

\item{nb_cl}{number of cores to use. (default \code{nb_cl = parallel::detectCores()-2})}

\item{nb_row_split}{row number for split files}

\item{returnOutputFileName}{wheter to return the new fragment file name}

\item{prefixMC}{prefix of metacell name in membership}
}
\description{
This function transforms fragments file associating each fragment to a single-cell to a fragments file associating fragments to metacells
}
\examples{
\dontrun{

}
}
