% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bmcite_HumanHTO7}
\alias{bmcite_HumanHTO7}
\title{Subset of the bone marrow CITE-seq dataset}
\format{
Seurat object
}
\source{

}
\usage{
bmcite_HumanHTO7
}
\description{
CITE-seq data of one tube (HumanHTO7) of BM cells from [Stuart T, Butler A, et al. 2019](https://doi.org/10.1016/j.cell.2019.05.031).
}
\details{
Data available as a seurat object in the SeuratData package' [GitHub](https://github.com/satijalab/seurat-data).
Data are derived from GEO: GSE128639
}
\keyword{datasets}
