% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_knn_graph.R
\name{build_knn_graph_nn2}
\alias{build_knn_graph_nn2}
\title{Build kNN graph using RANN::nn2
(used in \code{"build_knn_graph"})}
\usage{
build_knn_graph_nn2(X, k = min(5, ncol(X)))
}
\arguments{
\item{X}{matrix of coordinates (rows are samples and cols are coordinates)}

\item{k}{kNN parameter}

\item{return_neighbors_order}{whether return order of neighbors}
}
\value{
a list with components
\itemize{
  \item graph.knn - igraph object
}
}
\description{
Build kNN graph using RANN::nn2
(used in \code{"build_knn_graph"})
}
