% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SuperCell_for_Seurat.R
\name{DimPlotSC}
\alias{DimPlotSC}
\title{DimPlot of metacells in single cell space}
\usage{
DimPlotSC(
  seurat,
  seurat.mc,
  metacell.col = NULL,
  sc.col = NULL,
  dims = c(1, 2),
  reduction = "wnn.umap",
  mc.color = NULL,
  sc.color = NULL,
  alpha = 1,
  pt_size = 0,
  metric = "size",
  continuous_metric = F
)
}
\arguments{
\item{seurat}{the original seurat object at the single cell level with the computed dimension reduction}

\item{seurat.mc}{a seurat object at the metacell level}

\item{metacell.col}{name of metacell col in Seurat object (no need to use with default metacell obtained with \link{SCimplify_for_Seurat})}

\item{sc.col}{name of single-cell col in Seurat object}

\item{dims}{numerical vector of 2 dimension components to plot}

\item{reduction}{computed dimension reduction in the seurat object to use}

\item{mc.color}{colors for metacell idents}

\item{sc.color}{colors for single-cell idents}

\item{alpha}{transparency value for the single-cell points}

\item{pt_size}{size the single-cell points}

\item{continuous_metric}{bolean indicating if the metric variable is continuous or not. If TRUE a continuous color scale will be used}
}
\value{
a plot of metacells in single cell space
}
\description{
Plots metacell in single cell space (pca, umap,..) taking the average coordinates of single cells in each metacell
}
