% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SuperCell_Plots_with_Seurat.R
\name{DotPlot.SuperCell}
\alias{DotPlot.SuperCell}
\title{Dot plot of metacells
Take into account metacell sizes using ExpandMetacellSeuratAssay5
Same parameter as the original DotPlot Seurat function}
\usage{
DotPlot.SuperCell(
  object,
  features,
  assay = NULL,
  group.by = NULL,
  split.by = NULL,
  ...
)
}
\arguments{
\item{object}{A Seurat metacell object}

\item{features}{Input vector of features, or named list of feature vectors
if feature-grouped panels are desired (replicates the functionality of the
old SplitDotPlotGG)}

\item{assay}{Name of assay to use, defaults to the active assay}

\item{group.by}{Factor to group the cells by}

\item{split.by}{A factor in object metadata to split the plot by, pass 'ident'
to split by cell identity
see \code{\link[Seurat]{FetchData}} for more details}
}
\value{
As Seurat DotPLot, a ggplot object
}
\description{
Dot plot of metacells
Take into account metacell sizes using ExpandMetacellSeuratAssay5
Same parameter as the original DotPlot Seurat function
}
