% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SuperCell_Plots_with_Seurat.R
\name{ExpandMetacellSeuratAssay5}
\alias{ExpandMetacellSeuratAssay5}
\title{Expand metacells to single cells}
\usage{
ExpandMetacellSeuratAssay5(
  object,
  features,
  assay = "RNA",
  meta.data.vars = NULL
)
}
\arguments{
\item{object}{Seurat metacell object}
}
\value{
A seurat single cell object of expanded metacells
}
\description{
Given size of metacells create a single cell seurat object
in which each metacell of size s is expanded (same data and metadata)
to s single cells.
}
