% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SuperCell_Plots_with_Seurat.R
\name{FeatureFeaturePlot.SuperCell}
\alias{FeatureFeaturePlot.SuperCell}
\title{FeatureFeaturePlot from SuperCell seurat object}
\usage{
FeatureFeaturePlot.SuperCell(
  seurat.obj,
  feature.x,
  feature.y,
  cluster = NULL,
  assays = c("RNA", "RNA"),
  norm.margin.1 = 1,
  normalization.method.1 = "LogNormalize",
  norm.margin.2 = 1,
  normalization.method.2 = "LogNormalize",
  is.normalized = F,
  plot = F,
  color.use = NULL,
  use.size = T,
  add.key = T,
  ...
)
}
\arguments{
\item{seurat.obj}{A Seurat object or Seurat metacell object}

\item{feature.x}{Vector of features from the first assay to represent on the x axis}

\item{feature.y}{Vector of features from the second assay to represent on the y axis}

\item{cluster}{Names of the meta.data column to use to color the cells or metacells}

\item{assays}{(default is c("RNA", "RNA")) Vector of 2 assays from which `feature.x` and `feature.y` will be selected}

\item{norm.margin.1}{(default is 1) If performing CLR normalization for the first assay in `assays`, normalize across features (1) or cells (2)}

\item{normalization.method.1}{Method for normalization for the first assay in `assays`}

\item{norm.margin.2}{(default is 1) If performing CLR normalization for the second assay in `assays`, normalize across features (1) or cells (2)}

\item{normalization.method.2}{Method for normalization for the second assay in `assays`}

\item{is.normalized}{(default is FALSE) Bolean indicating whether to normalize the assays in `assays`}

\item{plot}{(default is FALSE) Bolean indicating whether to generate and save the scatter plots or only retrieve correlation values}

\item{use.size}{Consider metacell sizes in the scatter plot (dot sizes) and in the correlation test between the features in feature.x and feature.y}
}
\value{
A list of SingleCorPlot scatter plot for single-cells or metacells
}
\description{
FeatureFeaturePlot function using metacell sizes
}
