% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SuperCell_DE_with_Seurat.R
\name{FindAllMarkers.SuperCell}
\alias{FindAllMarkers.SuperCell}
\title{Gene expression markers for all identity classes}
\usage{
FindAllMarkers.SuperCell(
  object,
  assay = NULL,
  features = NULL,
  logfc.threshold = 0.1,
  test.use = "survey_weighted_t",
  slot = "data",
  min.pct = 0.01,
  min.diff.pct = -Inf,
  node = NULL,
  verbose = TRUE,
  only.pos = FALSE,
  max.cells.per.ident = Inf,
  random.seed = 1,
  latent.vars = NULL,
  min.cells.feature = 3,
  min.cells.group = 3,
  mean.fxn = NULL,
  fc.name = NULL,
  base = 2,
  return.thresh = 0.01,
  densify = FALSE,
  ...
)
}
\arguments{
\item{object}{Seurat metacell object with a metadata size column}

\item{assay}{Assay to use in differential expression testing}

\item{features}{Genes to test. Default is to use all genes}

\item{logfc.threshold}{Limit testing to genes which show, on average, at least
X-fold difference (log-scale) between the two groups of cells. Default is 0.1
Increasing logfc.threshold speeds up the function, but can miss weaker signals.
If the \code{slot} parameter is "scale.data" no filtering is performed.}

\item{test.use}{onky "weighted_t" is supported.}

\item{slot}{Slot to pull data from; note that if \code{test.use} is
"negbinom", "poisson", or "DESeq2", \code{slot} will be set to "counts"}

\item{min.pct}{only test genes that are detected in a minimum fraction of
min.pct cells in either of the two populations. Meant to speed up the function
by not testing genes that are very infrequently expressed. Default is 0.01}

\item{min.diff.pct}{only test genes that show a minimum difference in the
fraction of detection between the two groups. Set to -Inf by default}

\item{node}{A node to find markers for and all its children; requires
\code{\link[Seurat]{BuildClusterTree}} to have been run previously; replaces \code{FindAllMarkersNode}}

\item{verbose}{Print a progress bar once expression testing begins}

\item{only.pos}{Only return positive markers (FALSE by default)}

\item{max.cells.per.ident}{Down sample each identity class to a max number.
Default is no downsampling. Not activated by default (set to Inf)}

\item{random.seed}{Random seed for downsampling}

\item{latent.vars}{Variables to test, used only when \code{test.use} is one of
'LR', 'negbinom', 'poisson', or 'MAST'}

\item{min.cells.feature}{Minimum number of cells expressing the feature in at least one
of the two groups, currently only used for poisson and negative binomial tests}

\item{min.cells.group}{Minimum number of cells in one of the groups}

\item{mean.fxn}{Function to use for fold change or average difference calculation.
The default depends on the the value of \code{fc.slot}:
\itemize{
 \item{"counts"} : difference in the log of the mean counts, with pseudocount.
 \item{"data"} : difference in the log of the average exponentiated data, with pseudocount.
 This adjusts for differences in sequencing depth between cells, and assumes that "data"
 has been log-normalized.
 \item{"scale.data"} : difference in the means of scale.data.
}}

\item{fc.name}{Name of the fold change, average difference, or custom function column
in the output data.frame. If NULL, the fold change column will be named
according to the logarithm base (eg, "avg_log2FC"), or if using the scale.data
slot "avg_diff".}

\item{base}{The base with respect to which logarithms are computed.}

\item{return.thresh}{Only return markers that have a p-value < return.thresh, or a power > return.thresh (if the test is ROC)}

\item{densify}{Convert the sparse matrix to a dense form before running the 
DE test. This can provide speedups but might require higher memory; default is FALSE}

\item{...}{Arguments passed to other methods and to specific DE methods}
}
\value{
data.frame with a ranked list of putative markers as rows, and associated statistics as columns (p-values, t statistic, degree of freedom (df) of the weighted t-test). 
avg_logFC, pct.1, pc.2 are computed using Seurat v5 approach (pseudocounts added at the group level) taking into account metacell sizes.
}
\description{
Same as Seurat FindAllMarkers function taking into account metacell size.
Only weighted t test using weights::wtd.t.test funnction is currently supported
}
