% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SuperCell_DE_with_Seurat.R
\name{FindMarkers.SuperCell}
\alias{FindMarkers.SuperCell}
\alias{FindMarkers.SuperCell.default}
\alias{FindMarkers.SuperCell.Assay}
\alias{FindMarkers.SuperCell.Seurat}
\title{Find Markers of metacells}
\usage{
FindMarkers.SuperCell(object, ...)

\method{FindMarkers.SuperCell}{default}(
  object,
  slot = "data",
  cells.1 = NULL,
  cells.2 = NULL,
  weights.1 = NULL,
  weights.2 = NULL,
  features = NULL,
  logfc.threshold = 0.1,
  test.use = "survey_weighted_t",
  min.pct = 0.01,
  min.diff.pct = -Inf,
  verbose = TRUE,
  only.pos = FALSE,
  max.cells.per.ident = Inf,
  random.seed = 1,
  latent.vars = NULL,
  min.cells.feature = 3,
  min.cells.group = 3,
  fc.results = NULL,
  densify = FALSE,
  ...
)

\method{FindMarkers.SuperCell}{Assay}(
  object,
  slot = "data",
  cells.1 = NULL,
  cells.2 = NULL,
  weights.1 = NULL,
  weights.2 = NULL,
  features = NULL,
  test.use = "survey_weighted_t",
  fc.slot = "data",
  pseudocount.use = 1,
  norm.method = NULL,
  mean.fxn = NULL,
  fc.name = NULL,
  base = 2,
  ...
)

\method{FindMarkers.SuperCell}{Seurat}(
  object,
  ident.1 = NULL,
  ident.2 = NULL,
  latent.vars = NULL,
  group.by = NULL,
  subset.ident = NULL,
  assay = NULL,
  reduction = NULL,
  ...
)
}
\arguments{
\item{object}{Seurat metacell object}

\item{...}{Arguments passed to other methods and to specific DE methods}

\item{slot}{Slot to pull data from; note that if \code{test.use} is
"negbinom", "poisson", or "DESeq2", \code{slot} will be set to "counts"}

\item{cells.1}{Vector of cell names belonging to group 1}

\item{cells.2}{Vector of cell names belonging to group 2}

\item{features}{Genes to test. Default is to use all genes}

\item{logfc.threshold}{Limit testing to genes which show, on average, at least
X-fold difference (log-scale) between the two groups of cells. Default is 0.1
Increasing logfc.threshold speeds up the function, but can miss weaker signals.
If the \code{slot} parameter is "scale.data" no filtering is performed.}

\item{test.use}{onky "weighted_t" is supported.}

\item{min.pct}{only test genes that are detected in a minimum fraction of
min.pct cells in either of the two populations. Meant to speed up the function
by not testing genes that are very infrequently expressed. Default is 0.01}

\item{min.diff.pct}{only test genes that show a minimum difference in the
fraction of detection between the two groups. Set to -Inf by default}

\item{verbose}{Print a progress bar once expression testing begins}

\item{only.pos}{Only return positive markers (FALSE by default)}

\item{max.cells.per.ident}{Down sample each identity class to a max number.
Default is no downsampling. Not activated by default (set to Inf)}

\item{random.seed}{Random seed for downsampling}

\item{latent.vars}{Variables to test, used only when \code{test.use} is one of
'LR', 'negbinom', 'poisson', or 'MAST'}

\item{min.cells.feature}{Minimum number of cells expressing the feature in at least one
of the two groups, currently only used for poisson and negative binomial tests}

\item{min.cells.group}{Minimum number of cells in one of the groups}

\item{fc.results}{data.frame from FoldChange}

\item{densify}{Convert the sparse matrix to a dense form before running the 
DE test. This can provide speedups but might require higher memory; default is FALSE}

\item{fc.slot}{Slot used to calculate fold-change - will also affect the 
default for \code{mean.fxn}, see below for more details.}

\item{pseudocount.use}{Pseudocount to add to averaged expression values when
calculating logFC. 1 by default.}

\item{norm.method}{Normalization method for fold change calculation when
\code{slot} is \dQuote{\code{data}}}

\item{mean.fxn}{Function to use for fold change or average difference calculation.
The default depends on the the value of \code{fc.slot}:
\itemize{
 \item{"counts"} : difference in the log of the mean counts, with pseudocount.
 \item{"data"} : difference in the log of the average exponentiated data, with pseudocount.
 This adjusts for differences in sequencing depth between cells, and assumes that "data"
 has been log-normalized.
 \item{"scale.data"} : difference in the means of scale.data.
}}

\item{fc.name}{Name of the fold change, average difference, or custom function column
in the output data.frame. If NULL, the fold change column will be named
according to the logarithm base (eg, "avg_log2FC"), or if using the scale.data
slot "avg_diff".}

\item{base}{The base with respect to which logarithms are computed.}

\item{ident.1}{Identity class to define markers for; pass an object of class
\code{phylo} or 'clustertree' to find markers for a node in a cluster tree;
passing 'clustertree' requires \code{\link[Seurat]{BuildClusterTree}} to have been run}

\item{ident.2}{A second identity class for comparison; if \code{NULL},
use all other cells for comparison; if an object of class \code{phylo} or
'clustertree' is passed to \code{ident.1}, must pass a node to find markers for}

\item{group.by}{Regroup cells into a different identity class prior to 
performing differential expression (see example); \code{"ident"} to use Idents}

\item{subset.ident}{Subset a particular identity class prior to regrouping. 
Only relevant if group.by is set (see example)}

\item{assay}{Assay to use in differential expression testing}

\item{reduction}{Reduction to use in differential expression testing - will 
test for DE on cell embeddings}
}
\value{
data.frame with a ranked list of putative markers as rows, and associated statistics as columns (p-values, t statistic, degree of freedom (df) of the weighted t-test). 
avg_logFC, pct.1, pc.2 are computed using Seurat v5 approach (pseudocounts added at the group level) taking into account metacell sizes.

data.frame with a ranked list of putative markers as rows, and associated statistics as columns (p-values, t statistic, degree of freedom (df) of the weighted t-test). 
avg_logFC, pct.1, pc.2 are computed using Seurat v5 approach (pseudocounts added at the group level) taking into account metacell sizes.
}
\description{
Same as Seurat FindMarkers function taking into account metacell size.
Only weighted t test using weights::wtd.t.test funnction is currently supported

Same as Seurat FindMarkers function taking into account metacell size.
Only weighted t test using weights::wtd.t.test funnction is currently supported
}
\concept{differential_expression}
