% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SuperCell_FindMultimodalMarkers.R
\name{FindMultimodalMarkers.SuperCell}
\alias{FindMultimodalMarkers.SuperCell}
\title{Multimodal markers for all identity classes}
\usage{
FindMultimodalMarkers.SuperCell(
  seurat.obj,
  group.by = "celltype",
  assay1 = "RNA",
  assay2 = "chromvar",
  min.cells.feature = 0,
  min.cells.group = 0,
  min.pct = 0.01,
  padj.cutoff = 0.05,
  base = 2,
  only.pos = T,
  return.thresh = 1,
  logfc.threshold1 = 0.1,
  logfc.threshold2 = 0.1,
  test.use = "survey_weighted_t",
  fc.name1 = "avg_log2FC",
  fc.name2 = "avg_diff",
  mean.fxn1 = NULL,
  mean.fxn2 = weighted.mean.fxn,
  ...
)
}
\arguments{
\item{assay1}{Seurat assay to consider as the first modality}

\item{assay2}{Seurat assay to consider as the second modality}

\item{min.cells.feature}{Minimum number of cells expressing the feature in at least one
of the two groups, currently only used for poisson and negative binomial tests}

\item{min.cells.group}{Minimum number of cells in one of the groups}

\item{min.pct}{only test genes that are detected in a minimum fraction of
min.pct cells in either of the two populations. Meant to speed up the function
by not testing genes that are very infrequently expressed. Default is 0.01}

\item{base}{The base with respect to which logarithms are computed.}

\item{only.pos}{Only return positive markers (FALSE by default)}

\item{return.thresh}{Only return markers that have a p-value < return.thresh, or a power > return.thresh (if the test is ROC)}

\item{logfc.threshold1}{logfc.threshold parameter that FindAllMarkers.SuperCell will consider for the first modality}

\item{logfc.threshold2}{logfc.threshold parameter that FindAllMarkers.SuperCell will consider for the second modality}

\item{test.use}{(default is "survey_weighted_t") Test to use in FindAllMarkers.SuperCell}

\item{fc.name1}{fc.name parameter that FindAllMarkers.SuperCell will consider for the first modality}

\item{fc.name2}{fc.name parameter that FindAllMarkers.SuperCell will consider for the second modality}

\item{mean.fxn1}{mean.fxn parameter that FindAllMarkers.SuperCell will consider for the first modality}

\item{mean.fxn2}{mean.fxn parameter that FindAllMarkers.SuperCell will consider for the second modality}

\item{...}{Arguments passed to other methods and to specific DE methods}

\item{object}{Seurat metacell object with a metadata size column}
}
\value{
A list of 3 data.frames: (i) a data frame containing a ranked list of putative multimodal markers as rows, and associated statistics as columns (p-values, t statistic, degree of freedom (df) of the weighted t-test).
avg_logFC, pct.1, pc.2 are computed taking into account metacell sizes, (ii) a data frame containing the output of FindAllMarkers.SuperCell run on the first modality, and (iii) a data frame containing the output of FindAllMarkers.SuperCell run onthe second modality.
}
\description{
Function identifying markers in 2 modalities using FindAllMarkers.SuperCell
}
