% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_knn_graph.R
\name{knn_graph_from_dist}
\alias{knn_graph_from_dist}
\title{Build kNN graph from distance
(used in \code{"build_knn_graph"})}
\usage{
knn_graph_from_dist(D, k = 5, return_neighbors_order = T, mode = "all")
}
\arguments{
\item{D}{dist matrix or dist object (preferentially)}

\item{k}{kNN parameter}

\item{return_neighbors_order}{whether return order of neighbors (not available for nn2 option)}

\item{mode}{mode of \link[igraph]{graph_from_adj_list} ('all' -- undirected graph, 'out' -- directed graph)}
}
\value{
a list with components
\itemize{
  \item graph.knn - igraph object
  \item order - Nxk matrix with indices of k nearest neighbors ordered by relevance (from 1st to k-th)
}
}
\description{
Build kNN graph from distance
(used in \code{"build_knn_graph"})
}
