% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{metacell_anndata_2_supercell}
\alias{metacell_anndata_2_supercell}
\title{Convert Metacells (Metacell-2) to Super-cell like object}
\usage{
metacell_anndata_2_supercell(adata, obs.sc)
}
\arguments{
\item{adata}{anndata object of metacells (the output of \href{https://metacells.readthedocs.io/en/latest/_modules/metacells/pipeline/collect.html#collect_metacells}{\code{collect_metacells()}} function)}

\item{obs.sc}{a dataframe of the single-cell anndata object used to compute metacells (anndata after applying \href{https://metacells.readthedocs.io/en/latest/_modules/metacells/pipeline/divide_and_conquer.html#divide_and_conquer_pipeline}{\code{divide_and_conquer_pipeline()}} function)}
}
\value{
a list of super-cell like object (similar to the output of \link{SCimplify})
}
\description{
Convert Metacells (Metacell-2) to Super-cell like object
}
