% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metacell_ASW_CiLISI.R
\name{MetacellBatchSplitASW}
\alias{MetacellBatchSplitASW}
\title{MetacellBatchSplitASW}
\usage{
MetacellBatchSplitASW(
  sobj.mc,
  reduction.name,
  meta.data.sc,
  bio.label.col,
  batch.label.col,
  kept.labels = NULL,
  dims = NULL,
  return.res.all = F,
  min.obs = 10
)
}
\arguments{
\item{sobj.mc}{A metacell seurat object}

\item{reduction.name}{Metacell latent space to use to compute the silhouettes.}

\item{meta.data.sc}{A meta data dataframe of single cells..}

\item{kept.labels}{labels to keep to compute the silhouettes (default all). Useful to discard annotated cells.}

\item{label.col}{column name of bio labels in metacell meta data .}

\item{batch.col}{column name of batch labels in metacell meta data to use to compute the silhouettes.}

\item{return.all.res}{dimensions to use to compute the silhouettes (default all computed present in the Seurat object).}
}
\value{
A vector containing the compactness of each metacell.
}
\description{
\code{MetacellBatchSplitASW} 
Compute weigted silhouette of batch labels of the metacells after splitting. (relie on Seurat R packages). 
Compute weighted average silhouette for batch label in each cell type and take the mean (based on batch_ASW proposed by Luecken et al 2022)
}
\examples{
batch <- MetacellBatchSplitASW(sobj.mc = sobj.mc,
                         reduction.name = "integrated.rpca",
                         meta.data.sc = meta.data.sc,
                         dims = 1:50,
                         batch.label.col = "batch",
                         label.col = "label")
}
