% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metacell_ASW_CiLISI.R
\name{MetacellBioSingleASW}
\alias{MetacellBioSingleASW}
\title{MetacellBioSingleASW}
\usage{
MetacellBioSingleASW(
  sobj.mc,
  meta.data.sc,
  reduction.name,
  label.col,
  return.all.res = FALSE,
  dims = NULL,
  kept.labels = NULL
)
}
\arguments{
\item{sobj.mc}{A metacell suerat object}

\item{meta.data.sc}{A meta data dataframe of single cells..}

\item{reduction.name}{Metacell latent space to use to compute the silhouettes.}

\item{label.col}{label column in metacell meta data to use to compute the silhouettes.}

\item{return.all.res}{dimensions to use to compute the silhouettes (default all computed present in the Seurat object).}

\item{kept.labels}{labels to keep to compute the silhouettes (default all). Useful to discard annotated cells.}
}
\value{
A vector containing the compactness of each metacell.
}
\description{
\code{MetacellBioSingleASW} 
Compute silhouette of bio labels of the metacells at the single cell levels. (relies on scIntegrationMetrics and Seurat R packages).
}
\examples{
bioASW <- MetacellBioSingleASW(sobj.mc = sobj.mc,
                         reduction.name = "integrated.rpca",
                         meta.data.sc = meta.data.sc,
                         dims = 1:50,
                         label.col = "label")
}
