% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetacellExpression.R
\name{MetacellExpression}
\alias{MetacellExpression}
\title{MetacellExpression}
\usage{
MetacellExpression(
  object,
  pb.method = "aggregate",
  assays = NULL,
  features = NULL,
  return.seurat = TRUE,
  group.by = "ident",
  add.ident = NULL,
  layer = "counts",
  verbose = TRUE,
  ...
)
}
\arguments{
\item{object}{A Seurat single-cell object. It has to be preprocessed (eg. latent space computed) for the assay(s) used to identify metacells}
}
\value{
A Seurat object with aggregated data
}
\description{
\code{MetacellExpression} 
Compute metacells from a Seurat single-cell object.
}
