% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pancreas}
\alias{pancreas}
\title{Pancreatic cell dataset}
\format{
A list with spliced count matrix (emat), un-spliced count matrix (nmat) and metadata data frame (meta):
\describe{
  \item{emat}{spliced (exonic) count matrix}
  \item{nmat}{un-spliced (intronic) count matrix}
}
}
\source{
\url{https://scvelo.readthedocs.io/Pancreas.html}
}
\usage{
pancreas
}
\description{
Spliced and un-spliced scRNA-seq counts of 3696 pancreatic cells from  \href{https://dev.biologists.org/content/146/12/dev173849.abstract}{Bastidas-Ponce et al. (2018)}.
}
\keyword{datasets}
