% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singlecell_mixing.R
\name{sc_mixing_score}
\alias{sc_mixing_score}
\title{Compute mixing of single-cells within supercell}
\usage{
sc_mixing_score(SC, clusters)
}
\arguments{
\item{SC}{super-cell object (output of \code{\link{SCimplify}} function)}

\item{clusters}{vector of clustering assignment (reference assignment)}
}
\value{
a vector of single-cell mixing within super-cell it belongs to, which is defined as:
1 - proportion of cells of the same annotation (e.g., cell type) within the same super-cell
With 0 meaning that super-cell consists of single cells from one cluster (reference assignment) and higher values correspond to higher cell type mixing within super-cell
}
\description{
Compute mixing of single-cells within supercell
}
