% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCimplify.R
\name{SCimplify}
\alias{SCimplify}
\title{Detection of super-cells}
\usage{
SCimplify(
  X,
  genes.use = NULL,
  genes.exclude = NULL,
  n.var.genes = min(1000, nrow(X)),
  k.knn = 5,
  gamma = 10,
  do.scale = TRUE,
  n.pc = 10,
  fast.pca = TRUE,
  do.approx = FALSE,
  approx.N = 20000,
  use.nn2 = TRUE,
  seed = 12345,
  igraph.clustering = c("walktrap", "louvain"),
  return.singlecell.NW = TRUE,
  return.hierarchical.structure = TRUE,
  block.size = 10000
)
}
\arguments{
\item{X}{gene expression matrix with rows to be genes and cols to be cells}

\item{genes.use}{a vector of genes used to compute PCA}

\item{genes.exclude}{a vector of genes to be excluded when computing PCA}

\item{n.var.genes}{if \code{"genes.use"} is not provided, \code{"n.var.genes"} genes with the largest variation are used}

\item{k.knn}{parameter to compute single-cell kNN network}

\item{gamma}{graining level of data (proportion of number of single cells in the initial dataset to the number of super-cells in the final dataset)}

\item{do.scale}{whether to scale gene expression matrix to compure PCA}

\item{n.pc}{number of principal components to use for construction of single-cell kNN network}

\item{fast.pca}{use irlba::irlba as a faster version of prcomp (one used in Seurat package)}

\item{do.approx}{compute approximate kNN in case of a large dataset (>20000)}

\item{approx.N}{number of cells to subsample for an approximate approach}

\item{seed}{seed to use to subsample cells for an approximate approach}

\item{igraph.clustering}{clustering method to identify super-cells (available methods "walktrap" (default) and "louvain" (not recommended, gamma is ignored)).}

\item{return.singlecell.NW}{whether return single-cell network (which consists of approx.N if \code{"do.approx"} or all cells otherwise)}

\item{return.hierarchical.structure}{whether return hierarchical structure of super-cell}
}
\value{
a list with components
\itemize{
  \item graph.supercells - igraph object of a simplified network (number of nodes corresponds to number of super-cells)
  \item membership - assigmnent of each single cell to a particular super-cell
  \item graph.singlecells - igraph object (kNN network) of single-cell data
  \item supercell_size - size of super-cells
}
}
\description{
This function detects super-cells from single-cell gene expression matrix
}
