% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCimplify_for_velocity.R
\name{SCimplify_for_velocity}
\alias{SCimplify_for_velocity}
\title{Construct super-cells from spliced and un-spliced matrices}
\usage{
SCimplify_for_velocity(emat, nmat, gamma = NULL, membership = NULL, ...)
}
\arguments{
\item{emat}{spliced (exonic) count matrix}

\item{nmat}{unspliced (nascent) count matrix}

\item{gamma}{graining level of data (proportion of number of single cells in the initial dataset to the number of super-cells in the final dataset)}

\item{membership}{metacell membership vector (if provided, will be used for \code{emat}, \code{nmat} metacell matrices averaging)}

\item{...}{other parameters from \link{SCimplify}}
}
\value{
list containing vector of membership, spliced count and un-spliced count matrices
}
\description{
Construct super-cells from spliced and un-spliced matrices
}
