% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCimplify.R
\name{SCimplify_from_knn_graph}
\alias{SCimplify_from_knn_graph}
\title{Detection of metacells with the SuperCell approach from precomputed kNN graph}
\usage{
SCimplify_from_knn_graph(
  sc.nw,
  cell.annotation = NULL,
  cell.split.condition = NULL,
  gamma = 10,
  seed = 12345,
  igraph.clustering = c("walktrap", "louvain"),
  return.singlecell.NW = TRUE,
  return.hierarchical.structure = TRUE,
  ...
)
}
\arguments{
\item{sc.nw}{precomputed kNN graph (igraph object) with cells as nodes}

\item{cell.annotation}{a vector of cell type annotation, if provided, metacells that contain single cells of different cell type annotation will be split in multiple pure metacell (may result in slightly larger number of metacells than expected with a given gamma)}

\item{cell.split.condition}{a vector of cell conditions that must not be mixed in one metacell. If provided, metacells will be split in condition-pure metacell (may result in significantly(!) larger number of metacells than expected)}

\item{gamma}{graining level of data (proportion of number of single cells in the initial dataset to the number of metacells in the final dataset)}

\item{seed}{seed to use (not used in this function but kept for compatibility)}

\item{igraph.clustering}{clustering method to identify metacells (available methods "walktrap" (default) and "louvain" (not recommended, gamma is ignored)).}

\item{return.singlecell.NW}{whether return single-cell network}

\item{return.hierarchical.structure}{whether return hierarchical structure of metacell}

\item{...}{additional parameters (not used but kept for compatibility)}
}
\value{
a list with components
\itemize{
  \item graph.supercells - igraph object of a simplified network (number of nodes corresponds to number of metacells)
  \item membership - assignment of each single cell to a particular metacell
  \item graph.singlecells - igraph object (kNN network) of single-cell data
  \item supercell_size - size of metacells (former super-cells)
  \item gamma - requested graining level
  \item N.SC - number of obtained metacells
  \item genes.use - used genes (NA due to low-dim representation)
  \item sc.cell.annotation. - single-cell cell type annotation (if provided)
  \item sc.cell.split.condition. - single-cell split condition (if provided)
  \item SC.cell.annotation. - super-cell cell type annotation (if was provided for single cells)
  \item SC.cell.split.condition. - super-cell split condition (if was provided for single cells)

}
}
\description{
This function detects metacells from a precomputed kNN graph
}
