% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SuperCell_for_Seurat.R
\name{supercell_FeatureFeaturePlot_Seurat}
\alias{supercell_FeatureFeaturePlot_Seurat}
\title{Feature-feature correlation plot with a Seurat object}
\usage{
supercell_FeatureFeaturePlot_Seurat(
  seurat.mc,
  feature_x,
  feature_y,
  method = c("pearson", "kendall", "spearman"),
  assays = c("RNA", "ADT"),
  cluster = "celltype.l2",
  is.normalized = F,
  plot = T,
  color.use = NULL,
  use.size = T
)
}
\arguments{
\item{seurat.mc}{a seurat object}

\item{feature_x}{feature or vector of features (if vector, has to be the same lenght as feature_y)}

\item{feature_y}{feature or vector of features (if vector, has to be the same lenght as feature_x)}

\item{method}{a character string indicating which correlation coefficient is to be used for the test. One of "pearson", "kendall", or "spearman", can be abbreviated.}

\item{assays}{Assays to use in a vector (c(Assay_for_feature_x,Assay_for_feature_y))}

\item{plot}{if FALSE do not plot and only return the computed statistics}

\item{color.use}{colors for idents}

\item{use.size}{whether to compute weighted correlation by size (required that the seurat object has a size column in its metadata)}

\item{clusters}{seurat metadata column with clustering information}

\item{is.normalised}{if FALSE ADT assay will be normalized with "CLR" and RNA assay with a "LogNormalize"}
}
\value{
a data.frame with correlation coefficient for each feature couples
}
\description{
Plots feature-feature expression and computes their correlation
}
