% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supercell_GE.R
\name{supercell_GE}
\alias{supercell_GE}
\title{Simplification of scRNA-seq dataset}
\usage{
supercell_GE(ge, groups, weights = NULL, do.median.norm = FALSE)
}
\arguments{
\item{ge}{gene expression matrix (or any coordinate matrix) with genes as rows and cells as cols}

\item{groups}{vector of membership (assignment of single-cell to super-cells)}

\item{weights}{vector of a cell weight (NULL by default), used for computing average gene expression withing cluster of super-cells}

\item{do.median.norm}{whether to normalize}
}
\value{
a matrix of simplified (averaged withing groups) data with ncol equal to number of groups and nrows as in the initial dataset
}
\description{
This function converts gene-expression matrix of single-cell data into a gene expression
matrix of super-cells
}
