% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supercell_GeneGenePlot.R
\name{supercell_GeneGenePlot_single}
\alias{supercell_GeneGenePlot_single}
\title{Plot  Gene-gene correlation plot for 1 feature}
\usage{
supercell_GeneGenePlot_single(
  ge_x,
  ge_y,
  gene_x_name,
  gene_y_name,
  supercell_size = NULL,
  clusters = NULL,
  color.use = NULL,
  x.max = NULL,
  y.max = NULL,
  pt.size = 1,
  alpha = 0.9
)
}
\arguments{
\item{ge_x}{first gene expression vector (same length as number of super-cells)}

\item{ge_y}{second gene expression vector (same length as number of super-cells)}

\item{gene_x_name}{name of gene x}

\item{gene_y_name}{name of gene y}

\item{supercell_size}{a vector with supercell size (ordered the same way as in \code{ge})}

\item{clusters}{a vector with clustering information (ordered the same way as in \code{ge})}

\item{color.use}{colors for idents}

\item{x.max}{max of x axis}

\item{y.max}{max of y axis}

\item{pt.size}{point size (0 by default)}

\item{alpha}{transparency of dots}
}
\description{
Used for \link{supercell_GeneGenePlot}
}
