% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supercell_merge.R
\name{supercell_merge}
\alias{supercell_merge}
\title{Merging independent SuperCell objects}
\usage{
supercell_merge(SCs)
}
\arguments{
\item{SCs}{list of SuperCell objects (resuls of \link[SCimplify])}
}
\value{
a list with components
\itemize{
  \item membership - assigmnent of each single cell to a particular metacell
  \item cell.ids - the original ids of single-cells
  \item supercell_size - size of metacells (former super-cells)
  \item gamma -  graining level of the merged object (astimated as an average size of metacells as the independent SuperCell objects might have different graining levels)
  \item N.SC - number of obtained metacells
}
}
\description{
This function merges independent SuperCell objects
}
