% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supercell_merge.R
\name{supercell_mergeGE}
\alias{supercell_mergeGE}
\title{Merging metacell gene expression matrices from several independent SuperCell objects}
\usage{
supercell_mergeGE(SC.GEs)
}
\arguments{
\item{SC.GEs}{list of metacell gene expression matricies (result of \link{supercell_GE} ), make sure the order of the gene expression metricies is the same as in the call of \link{supercell_merge}}
}
\value{
a merged matrix of gene expression
}
\description{
This function merges independent SuperCell objects
}
\examples{
\dontrun{
# see examples in \link{supercell_merge}
}
}
