% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supercell_coverage.R
\name{supercell_NonZero}
\alias{supercell_NonZero}
\title{Compute the sparsity of gene counts}
\usage{
supercell_NonZero(GE, membership, genes)
}
\arguments{
\item{genes}{the set of gene to inspect}
}
\description{
Compute the sparsity of gene counts
}
