% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supercell_plot_GE.R
\name{supercell_plot_GE}
\alias{supercell_plot_GE}
\title{Plot super-cell NW colored by an expression of a gene (gradient color)}
\usage{
supercell_plot_GE(
  SC.nw,
  ge,
  color.use = c("gray", "blue"),
  n.color.gradient = 10,
  lay.method = c("nicely", "fr", "components", "drl", "graphopt", "dh"),
  lay = NULL,
  alpha = alpha,
  seed = 12345,
  main = NA,
  do.frames = TRUE,
  do.extra.log.rescale = FALSE,
  log.base = 2,
  do.extra.sqtr.rescale = FALSE,
  frame.color = "black",
  weights = NULL,
  min.cell.size = 0,
  legend.side = 4,
  gene.name = NULL
)
}
\arguments{
\item{SC.nw}{a super-cell network (a field \code{supercell_network} of the output of \link{SCimplify})}

\item{ge}{a gene expression vector (same length as number of super-cells)}

\item{color.use}{colors of gradient}

\item{n.color.gradient}{number of bins of the gradient, default is 10}

\item{do.extra.log.rescale}{= FALSE,}

\item{legend.side}{a side parameter of \link[plotfunctions]{gradientLegend} fucntion (default is 4)}

\item{gene.name}{name of gene of for which gene expression is plotted}

\item{...}{rest of the parameters of \link{supercell_plot} function}
}
\value{
plot of a super-cell network with color representing an expression level
}
\description{
Plot super-cell NW colored by an expression of a gene (gradient color)
}
