% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supercell_tSNE.R
\name{supercell_plot_tSNE}
\alias{supercell_plot_tSNE}
\title{Plot super-cell tSNE
Plots super-cell tSNE (result of \link{supercell_tSNE})}
\usage{
supercell_plot_tSNE(
  SC,
  groups,
  tSNE_name = "SC_tSNE",
  color.use = NULL,
  asp = 1,
  alpha = 0.7,
  title = NULL,
  ...
)
}
\value{
\link[ggplot2]{ggplot}
}
\description{
Plot super-cell tSNE
Plots super-cell tSNE (result of \link{supercell_tSNE})
}
