% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supercell_UMAP.R
\name{supercell_plot_UMAP}
\alias{supercell_plot_UMAP}
\title{Plot super-cell UMAP
Plots super-cell UMAP (result of \link{supercell_UMAP})}
\usage{
supercell_plot_UMAP(
  SC,
  groups,
  UMAP_name = "SC_UMAP",
  color.use = NULL,
  asp = 1,
  alpha = 0.7,
  title = NULL,
  ...
)
}
\value{
\link[ggplot2]{ggplot}
}
\description{
Plot super-cell UMAP
Plots super-cell UMAP (result of \link{supercell_UMAP})
}
