% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supercell_purity.R
\name{supercell_purity_GE}
\alias{supercell_purity_GE}
\title{Mean transcriptional purity}
\usage{
supercell_purity_GE(GE, membership, exclude.single.cells = FALSE)
}
\description{
@param GE gene expression matrix
 @param membership membership vector (or a list membership vectors)
 @export
}
