% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supercell_rescale.R
\name{supercell_rescale}
\alias{supercell_rescale}
\title{Rescale supercell object}
\usage{
supercell_rescale(SC.object, gamma = NULL, N.SC = NULL)
}
\arguments{
\item{SC.object}{super-cell object (an output from \code{\link{SCimplify}} function)}

\item{gamma}{new grainig level (provide either \code{gamma} or \code{N.SC})}

\item{N.SC}{new number of super-cells (provide either \code{gamma} or \code{N.SC})}
}
\value{
the same object as \code{\link{SCimplify}}
}
\description{
This function recomputes super-cell structure at a different graining level (\code{gamma}) or
for a specific number of super-cells (\code{N.SC})
}
