% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supercell_silhouette.R
\name{supercell_silhouette}
\alias{supercell_silhouette}
\title{Compute Silhouette index accounting for samlpe size (super cells size) ###}
\usage{
supercell_silhouette(x, dist, supercell_size = NULL)
}
\arguments{
\item{x}{-- clustering}

\item{dist}{- distance among super-cells}

\item{supercell_size}{-- super-cell size}
}
\value{
silhouette result
}
\description{
Compute Silhouette index accounting for samlpe size (super cells size) ###
}
