% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supercell_tSNE.R
\name{supercell_tSNE}
\alias{supercell_tSNE}
\title{Compute tSNE of super-cells}
\usage{
supercell_tSNE(
  SC,
  PCA_name = "SC_PCA",
  n.comp = NULL,
  perplexity = 30,
  seed = 12345,
  ...
)
}
\arguments{
\item{SC}{super-cell structure (output of \link{SCimplify}) with a field \code{PCA_name} containig PCA result}

\item{PCA_name}{name of \code{SC} field containing result of \link{supercell_prcomp}}

\item{n.comp}{number of vector of principal components to use for computing tSNE}

\item{perplexity}{perplexity parameter (parameter of \link[Rtsne]{Rtsne})}

\item{seed}{random seed}

\item{...}{other parameters of \link[Rtsne]{Rtsne}}
}
\value{
\link[Rtsne]{Rtsne} result
}
\description{
Computes tSNE of super-cells
}
