% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supercell_VlnPlot.R
\name{supercell_VlnPlot_single}
\alias{supercell_VlnPlot_single}
\title{Plot  Violin plot for 1 feature}
\usage{
supercell_VlnPlot_single(
  ge1,
  supercell_size = NULL,
  clusters,
  feature = NULL,
  color.use = NULL,
  pt.size = 0,
  pch = "o",
  y.max = NULL,
  y.min = NULL,
  adjust = 1,
  angle.text.y = 90,
  angle.text.x = 45
)
}
\arguments{
\item{ge1}{a gene expression vector (same length as number of super-cells)}

\item{supercell_size}{a vector with supercell size (ordered the same way as in \code{ge})}

\item{clusters}{a vector with clustering information (ordered the same way as in \code{ge})}

\item{feature}{gene to plot}

\item{color.use}{colors for idents}

\item{pt.size}{point size (0 by default)}

\item{pch}{shape of jitter dots}

\item{y.max}{max of y axis}

\item{y.min}{min of y axis}

\item{adjust}{param of geom_violin}

\item{angle.text.y}{rotation of y text}

\item{angle.text.x}{rotation of x text}
}
\description{
Used for supercell_VlnPlot
}
