% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_fragment_file.R
\name{transform_fragment_file}
\alias{transform_fragment_file}
\title{Transform fragments file}
\usage{
transform_fragment_file(
  input_file,
  output_name,
  output_path,
  membership,
  n_skip = 0
)
}
\arguments{
\item{input_file}{path to original fragments file}

\item{output_name}{name of the supercell fragments file (whithout extension)}

\item{output_path}{path where the new fragments file will be saved}

\item{membership}{membership vector}

\item{n_skip}{number of lines to skip in fragments file, (by default 0)}
}
\description{
This function transforms fragments file associating each fragment to a single-cell to a fragments file associating fragments to metacells
}
\examples{
\dontrun{

}
}
