% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SuperCell_Plots_with_Seurat.R
\name{VlnPlot.SuperCell}
\alias{VlnPlot.SuperCell}
\title{Violin plot of metacells
Take into account metacell sizes using ExpandMetacellSeuratAssay5
Same parameter as the original VlnPlot Seurat function}
\usage{
VlnPlot.SuperCell(
  object,
  features,
  assay = NULL,
  group.by = NULL,
  split.by = NULL,
  ...
)
}
\arguments{
\item{object}{A Seurat metacell object}

\item{features}{Features to plot (gene expression, metrics, PC scores,
anything that can be retreived by FetchData)}

\item{assay}{Name of assay to use, defaults to the active assay}

\item{group.by}{Group (color) cells in different ways (for example, orig.ident)}

\item{split.by}{A factor in object metadata to split the plot by, pass 'ident'
to split by cell identity}
}
\value{
As Seurat VlnPLot, a patchworked ggplot object if combine = TRUE; otherwise, a list of ggplot objects
}
\description{
Violin plot of metacells
Take into account metacell sizes using ExpandMetacellSeuratAssay5
Same parameter as the original VlnPlot Seurat function
}
