% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SuperCell_DE_with_Seurat.R,
%   R/SuperCell_FindMultimodalMarkers.R
\name{weighted.mean.fxn}
\alias{weighted.mean.fxn}
\title{Weighted mean}
\usage{
\method{weighted.mean}{fxn}(x, weights)

\method{weighted.mean}{fxn}(x, weights)
}
\arguments{
\item{x}{vector of observation (eg. metacell counts for a gene)*}

\item{weights}{vector of weights, same length as x (eg. metacell sizes)}
}
\description{
Weighted mean

Weighted mean function
}
