% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supercell_estimate_velocity.R
\name{supercell_estimate_velocity}
\alias{supercell_estimate_velocity}
\title{Run RNAvelocity for super-cells (slightly modified from \link[velocyto.R]{gene.relative.velocity.estimates})
Not yet adjusted for super-cell size (not sample-weighted)}
\usage{
supercell_estimate_velocity(
  emat,
  nmat,
  smat = NULL,
  membership = NULL,
  supercell_size = NULL,
  do.run.avegaring = (ncol(emat) == length(membership)),
  kCells = 1,
  ...
)
}
\arguments{
\item{emat}{spliced (exonic) count matrix}

\item{nmat}{unspliced (nascent) count matrix}

\item{membership}{supercell membership ('membership' field of \link{SCimplify})}

\item{supercell_size}{a vector with supercell size (if emat and nmat provided at super-cell level)}

\item{...}{other parameters from \link[velocyto.R]{gene.relative.velocity.estimates}}
}
\value{
reslts of \link[velocyto.R]{gene.relative.velocity.estimates} plus
}
\description{
Run RNAvelocity for super-cells (slightly modified from \link[velocyto.R]{gene.relative.velocity.estimates})
Not yet adjusted for super-cell size (not sample-weighted)
}
