% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supercell_plot.R
\name{supercell_plot}
\alias{supercell_plot}
\title{Plot super-cell NW}
\usage{
supercell_plot(
  SC.nw,
  group = NULL,
  color.use = NULL,
  lay.method = c("nicely", "fr", "components"),
  lay = NULL,
  seed = 12345,
  main = NA,
  do.frames = TRUE,
  do.extra.log.rescale = FALSE,
  log.base = 2,
  do.extra.sqtr.rescale = FALSE,
  frame.color = "black",
  weights = NULL,
  min.cell.size = 0
)
}
\arguments{
\item{SC.nw}{a super-cell network (a field \code{supercell_network} of the output of \link{SCimplify})}

\item{group}{an assigment of super cell to any group (for ploting in different colors)}

\item{color.use}{colros to use for groups, if \code{NULL}, an automatic palette of colors will be applied}

\item{lay.method}{method to compute layout of the network (for the moment there several available: "nicely"
for \link[igraph]{layout_licely} and "fr" for \link[igraph]{layout_with_fr}, "components" for \link[igraph]{layout_components})}

\item{lay}{a particular layout of a graph to plot (in is not \code{NULL}, \code{lay.method} is ignored and new layout is not computed)}

\item{seed}{a random seed used to compute graph layout}

\item{main}{a title of a plot}

\item{do.frames}{whether to keep vertex.frames in the plot}
}
\value{
plot of a super-cell network
}
\description{
Plot super-cell NW
}
