% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supercell_tSNE.R
\name{supercell_plot_tSNE}
\alias{supercell_plot_tSNE}
\title{Plot super-cell tSNE
(Use \link{supercell_DimPlot} instead)
Plots super-cell tSNE (result of \link{supercell_tSNE})}
\usage{
supercell_plot_tSNE(
  SC,
  groups,
  tSNE_name = "SC_tSNE",
  color.use = NULL,
  asp = 1,
  alpha = 0.7,
  title = NULL
)
}
\arguments{
\item{SC}{super-cell structure (output of \link{SCimplify}) with a field \code{tSNE_name} containing tSNE result}

\item{groups}{coloring metacells by groups}

\item{tSNE_name}{the mane of the field containing tSNE result}

\item{color.use}{colors of groups}

\item{asp}{plot aspect ratio}

\item{alpha}{transparency of}

\item{title}{title of the plot}
}
\value{
\link[ggplot2]{ggplot}
}
\description{
Plot super-cell tSNE
(Use \link{supercell_DimPlot} instead)
Plots super-cell tSNE (result of \link{supercell_tSNE})
}
