% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supercell_UMAP.R
\name{supercell_plot_UMAP}
\alias{supercell_plot_UMAP}
\title{Plot super-cell UMAP
(Use \link{supercell_DimPlot} instead)
Plots super-cell UMAP (result of \link{supercell_UMAP})}
\usage{
supercell_plot_UMAP(
  SC,
  groups,
  UMAP_name = "SC_UMAP",
  color.use = NULL,
  asp = 1,
  alpha = 0.7,
  title = NULL
)
}
\arguments{
\item{SC}{super-cell structure (output of \link{SCimplify}) with a field \code{UMAP_name} containing UMAP result}

\item{groups}{coloring metacells by groups}

\item{UMAP_name}{the mane of the field containing UMAP result}

\item{color.use}{colors of groups}

\item{asp}{plot aspect ratio}

\item{alpha}{transparency of}

\item{title}{title of the plot}
}
\value{
\link[ggplot2]{ggplot}
}
\description{
Plot super-cell UMAP
(Use \link{supercell_DimPlot} instead)
Plots super-cell UMAP (result of \link{supercell_UMAP})
}
