% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supercell_purity.R
\name{supercell_purity_GE}
\alias{supercell_purity_GE}
\title{Purity based on transcriptionally}
\usage{
supercell_purity_GE(GE, membership)
}
\description{
@param GE gene expression matrix
 @param membership membership vector
 @export
}
